Don't escape path parameters.
Confluence Data Center has a bug where if you send url encoded slashes, it will return 400 bad request. This hasn't been fixed in at least 5 years: https://community.developer.atlassian.com/t/confluence-rest-api-and-group-name-with/31968/15

It turns out there is no way to list groups with a slash or list who is a member of those groups.

So instead of failing a sync due to a 400, we get a 404 and sync everything but groups that the API can't process.